pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Ed-Host <kreker783@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
main = "gendiff.scripts.main:main"


[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"
coverage = "^7.3.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
